Adds mip benchmark

+++ src/codeql/codeql.py
@@ -9,24 +9,25 @@ import time, shutil
 
 class CodeQL:
     @staticmethod
-    def analyze(app_path, queries, rerun=False, num_runs=1):
+    def analyze(app_path, queries, rerun_db_create, rerun_queries, num_runs=1):
+        
         db_create_time_list = []
         query_time_list = []
         for i in range(num_runs):
             logger.info(f"Running iteration {i+1}/{num_runs} of CodeQL analysis...")
 
             app_path = f"{pathlib.Path(app_path).resolve()}"
-            codeql_db_path = f"{app_path}/codeqldb/"
-            output_dir = f"{app_path}/output/"
-            if rerun:
-                if os.path.exists(codeql_db_path):
-                    logger.info("Deleting existing database...")
-                    shutil.rmtree(codeql_db_path, ignore_errors=False, onerror=None)
-                    logger.info("Existing database deleted")
-                if os.path.exists(output_dir):
-                    logger.info("Deleting output directory...")
-                    shutil.rmtree(output_dir, ignore_errors=False, onerror=None)
-                    logger.info("Existing output directory deleted")
+            growlithe_path = f"{app_path}/growlithe/"
+            codeql_db_path = f"{growlithe_path}/codeqldb/"
+            output_path = f"{growlithe_path}/output/"
+            if rerun_db_create and os.path.exists(codeql_db_path):
+                logger.info("Deleting existing database...")
+                shutil.rmtree(codeql_db_path, ignore_errors=False, onerror=None)
+                logger.info("Existing database deleted")
+            if rerun_queries and os.path.exists(output_path):
+                logger.info("Deleting output directory...")
+                shutil.rmtree(output_path, ignore_errors=False, onerror=None)
+                logger.info("Existing output directory deleted")
 
             if os.path.exists(codeql_db_path):
                 logger.info(
@@ -34,7 +35,7 @@ class CodeQL:
                 )
             else:
                 start_time = time.time()
-                CodeQL._create_database(app_path)
+                CodeQL._create_database(app_path, growlithe_path)
                 printPattern("*", 75)
                 db_create_time_list.append(time.time() - start_time)
                 logger.info(
@@ -42,13 +43,13 @@ class CodeQL:
                 )
                 printPattern("*", 75)
 
-            if os.path.exists(output_dir):
+            if os.path.exists(output_path):
                 logger.info(
-                    f"CodeQL analysis output already exists at {output_dir} - skipping analysis"
+                    f"CodeQL analysis output already exists at {output_path} - skipping analysis"
                 )
             else:
                 start_time = time.time()
-                CodeQL._analyze_functions(app_path, codeql_db_path, output_dir, queries)
+                CodeQL._analyze_functions(app_path, codeql_db_path, output_path, queries)
                 query_time_list.append(time.time() - start_time)
 
             logger.info(f"Iteration {i+1}/{num_runs} of CodeQL analysis complete")
@@ -61,27 +62,20 @@ class CodeQL:
         )
 
     @staticmethod
-    def _analyze_functions(app_path, codeql_db_path, output_dir, queries):
+    def _analyze_functions(app_path, codeql_db_path, output_path, queries):
         current_dir = pathlib.Path(__file__).parent.resolve()
 
-        if os.path.exists(output_dir):
+        if os.path.exists(output_path):
             logger.info("Deleting existing output files...")
-            for file in glob.glob(f"{output_dir}/*"):
+            for file in glob.glob(f"{output_path}/*"):
                 os.remove(file)
             logger.info("Existing output directory deleted")
         else:
             logger.info("Creating output directory...")
-            os.mkdir(output_dir)
+            os.mkdir(output_path)
             logger.info("Output directory created")
 
         functions = find_python_files(app_path)
-
-        # Convert the Python list to a string
-        array_string = ", ".join([f'"{item}"' for item in functions])
-        # Create the final string in the desired format
-        result_string = f"result = [{array_string}]"
-
-
         logger.info(f"Found {len(functions)} functions to analyze: ")
         logger.info(functions)
 
@@ -93,11 +87,16 @@ class CodeQL:
         array_string = ",\n\t\t".join([f'"{item}"' for item in functions])
         # Create the final string in the desired format
         result_string = f"result = [{array_string}]"
-        pattern = re.compile(r"result\s*in\s*\[.*?\]", re.DOTALL)
+        pattern = re.compile(r"result\s*=\s*\[.*?\]", re.DOTALL)
+        if not re.search(pattern, config_template):
+            raise Exception("Could not find result variable in Config.qll")
+
         content = pattern.sub(result_string, config_template)
 
         with open(codeql_config_path, "w") as file:
             file.write(content)
+
+        logger.info(f"Rewritten {codeql_config_path} to add {result_string} ")
 
         for query_file in queries:
             printPattern("-", 75)
@@ -111,7 +110,7 @@ class CodeQL:
                     "analyze",
                     "-q",
                     "--output",
-                    f"{output_dir}/{query_file}.sarif",
+                    f"{output_path}/{query_file}.sarif",
                     "--format",
                     "sarifv2.1.0",
                     "--rerun",
@@ -134,10 +133,11 @@ class CodeQL:
             printPattern("*", 75)
 
     @staticmethod
-    def _create_database(app_path):
-        logger.info("Creating CodeQL database...")
+    def _create_database(app_path, growlithe_path):
+        printPattern("*", 75)
+        logger.info(f"Creating CodeQL database in {growlithe_path}")
         subprocess.run(
-            f"(cd {app_path} && codeql database create codeqldb --language=python --overwrite -j=0 -M=2048)",
+            f"(cd {growlithe_path} && codeql database create codeqldb --language=python --overwrite -j=0 -M=2048 -s={app_path})",
             shell=True,
             stdout=subprocess.DEVNULL,
         )
